Add ESP32 serial port mappings
This maps the logical serial port numbers to the internal ones used by the macro, while also making use of my previous patch that avoids having to edit the library directly. Instead, on Arduino, the user can simply leave macros unset unless they don't want to use software serial, and on ESP32 we throw a compile error if one of the hardware serial macros aren't set.

I opted to forego the option for hardware serial since picking which implementation of software serial for ESP32 to use is a big step. That task would be more suited to Robert himself.